# helpan-ai-openapi-v1.yaml
openapi: 3.1.0
info:
  title: Helpan AI Rail
  version: '1.0.0'
  summary: Agent rail for the Kirimon Market Ventures platform — orchestration runtime, delegated authority, consent surface, and audit for AI agents acting on behalf of users across Kipkiren Pay, Identiti, and Todoku.
  description: |
    Helpan AI is the **fourth platform rail** in the Kirimon Market Ventures (KMV) stack, alongside
    Identiti (identity), Kipkiren Pay (payment), and Todoku (comms). It provides the agent orchestration
    runtime, briefing storage, matching engine, delegated authority issuance/validation/revocation, the
    Helpan Console consent surface, audit log, and OAuth scope catalogue for third-party agents.

    Per design law (Helpan AI Instruction Pack §3) and the Agentic AI Signal Scan integration (7 May 2026):
    - Helpan AI does NOT hold funds, send messages, or store identity material. Every primitive that
      belongs to KP, Identiti, or Todoku is invoked via the relevant rail's API.
    - Every agent action carries a scoped, time-bounded, revocable **delegated authority token**.
    - Every issued delegated authority is visible in the Helpan Console with one-tap revocation.
    - Per-call validation of delegated authorities against this rail's `POST /authorities/{id}/validate`
      endpoint is mandatory for relying parties (KP, Todoku, consuming-app servers).
    - `initiated_by` and `actor` claims propagate cross-rail per Schema Appendix amendments §A on
...

...
  contact:
...
servers:
...
security:
  - HmacSha256: []
  - BearerCustomer: []
  - BearerService: []
paths:
...
  /authorities:
    post:
      tags: [Authorities]
      summary: Issue a delegated authority token (user-initiated, in-app)
      operationId: issueAuthority
      description: |
        Issues a scoped, time-bounded, revocable delegated authority token. The user is in-session
        and explicitly authorising the agent to act on a specified scope (data category, operation,
        spend limit, time window).

        For high-stakes scopes (money-touching, identity-sensitive), this endpoint requires a
        step-up token from Identiti in the request body. Per Confirmation Memo §5.9 and Identiti
        Schema Appendix Amendment §A, the step-up JWT carries the `actor` claim distinguishing
        subject from actor; this rail propagates the claim into the issued delegated authority's
        own JWT.
      parameters:
        - $ref: '#/components/parameters/XIdempotencyKey'
        - $ref: '#/components/parameters/Traceparent'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueAuthorityRequest'
      responses:
        '201':
          description: Authority issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorityResponse'
        '400':
          $ref: '#/components/responses/RequestInvalid'
        '401':
          description: Step-up required or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

    get:
      tags: [Authorities]
      summary: List active delegated authorities for the authenticated user (powers the Helpan Console)
      operationId: listAuthorities
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, expired, revoked]
        - name: agent_id
          in: query
          schema: { type: string }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        '200':
          description: Authorities list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorityListResponse'
  /authorities/{authority_id}:
    parameters:
      - $ref: '#/components/parameters/AuthorityId'
    get:
      tags: [Authorities]
      summary: Read a single authority detail
      operationId: getAuthority
      responses:
        '200':
          description: Authority detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorityResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /authorities/{authority_id}/revoke:
    parameters:
      - $ref: '#/components/parameters/AuthorityId'
    post:
      tags: [Authorities]
      summary: Revoke a delegated authority (immediate; propagates to relying parties)
      operationId: revokeAuthority
      description: |
        Revocation is **immediate**. The revocation propagates to relying parties via:
        1. Synchronous: future `POST /authorities/{id}/validate` calls return `revoked=true`.
        2. Asynchronous (v1.0): Kafka event `helpan.authority.events` with `AUTHORITY_REVOKED`.
        3. **CAEP push-revocation**: v1.1 roadmap (per Identiti Rail Contract Amendment §A.5).

        Relying parties (Kipkiren Pay, Todoku) must validate per call (60-second positive cache,
        **no negative cache**) so revocation is reflected within at most 60 seconds for poll-based
        validation. v1.1 CAEP cuts this to <5 seconds.
      parameters:
        - $ref: '#/components/parameters/XIdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeAuthorityRequest'
      responses:
        '200':
          description: Revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorityResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Already revoked or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /authorities/{authority_id}/validate:
    parameters:
      - $ref: '#/components/parameters/AuthorityId'
    post:
      tags: [Authorities]
      summary: Validate a delegated authority (relying-party endpoint)
      operationId: validateAuthority
      description: |
        Called by relying parties (Kipkiren Pay, Todoku) on every agent-dispatched call that
        carries the `X-Delegated-Authority` header. Returns the authority's current status,
        scope, and limits so the relying party can enforce.

        **Caching guidance:** 60-second positive cache (status='active') is permitted.
        **No negative caching** of revoked/expired status is permitted — relying parties must
        re-validate.

        **Authentication:** HMAC-SHA-256 with the relying-party rail's app credentials, scoped
        with `helpan:authority:validate`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateAuthorityRequest'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateAuthorityResponse'
        '401':
          $ref: '#/components/responses/AuthHmacInvalid'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================================================
  # ACTIONS — agent action dispatch
  # ==========================================================================
...
components:
...
  securitySchemes:
    HmacSha256:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Format: `Helpan-HMAC-SHA256 app_id={app_id}, signature={base64_sig}`.
        Canonical string (identical across all four rails per App Integration Guide §5.7):
        `{METHOD}\n{PATH_AND_QUERY}\n{CONTENT_TYPE}\n{X-Helpan-Timestamp}\n{SHA256_HEX(body)}`.
        mTLS additionally required at the transport layer.

    BearerCustomer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Identiti-issued customer JWT. Audience must include this rail's audience
        (`https://api.helpan.co.ke`). Carries `sub` (Account UUID), `tier`, `actor`,
        `initiated_by`, `auth_factors`, `exp`. Per Identiti Rail Contract Amendment §A.4:
        30-min standard / 5-min elevated TTL.

    BearerService:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Identiti-issued service JWT for service-to-service authentication
        (consuming-app servers calling Helpan AI on behalf of the user). 15-min standard /
        5-min elevated TTL per Identiti Rail Contract Amendment §A.4.

  # --------------------------------------------------------------------------
  # PARAMETERS
  # --------------------------------------------------------------------------
  parameters:
...
    XIdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: |
        UUIDv4 per logical operation. Apps SHOULD use the UUIDv5-from-business-op-id pattern
        per App Integration Guide §7.6. 24-hour idempotency window.
...
    Traceparent:
      name: Traceparent
      in: header
      required: false
      schema: { type: string, pattern: '^00-[0-9a-f]{32}-[0-9a-f]{16}-[0-9a-f]{2}$' }
      description: W3C Trace Context. Propagate unchanged across rails for cross-rail correlation.
...
    AuthorityId:
      name: authority_id
      in: path
      required: true
      schema: { type: string, pattern: '^daa_[0-9A-HJKMNP-TV-Z]{26}$' }
      description: Delegated-authority identifier (`daa_<ULID>`). Same value as the JWT `jti` claim.
...
  schemas:
...
    Meta:
      type: object
      required: [request_id, timestamp, schema_version]
      properties:
        request_id:
          type: string
          description: ULID
        timestamp:
          type: string
          format: date-time
        schema_version:
          type: string
          enum: ['1.0']
    SuccessEnvelope:
      type: object
      required: [ok, data, meta]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          description: Operation-specific payload
        meta:
          $ref: '#/components/schemas/Meta'
    ErrorEnvelope:
      type: object
      required: [ok, error, meta]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          $ref: '#/components/schemas/ErrorBody'
        meta:
          $ref: '#/components/schemas/Meta'
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code (e.g. `AUTH_HMAC_INVALID`)
        message:
          type: string
          description: Human-readable; not stable across versions
        field:
          type: string
          nullable: true
        detail:
          type: object
          additionalProperties: true
        documentation_url:
          type: string
          format: uri

    # ---- shared atoms ------------------------------------------------------
...
    AccountUuid:
      type: string
      pattern: '^acc_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
    KesMinorUnits:
      type: integer
      minimum: 0
      description: KES minor units (fils). Integer only. No floating-point. Per Reboot Pack §9.2.

    # ---- briefings ---------------------------------------------------------
...
    AuthorityScope:
      type: object
      required: [scope_id]
      properties:
        scope_id:
          type: string
          description: 'OAuth scope identifier (e.g. `kipkiren.write.payments`, `chapaa.mmf.rebalance`)'
        amount_limit_minor:
          $ref: '#/components/schemas/KesMinorUnits'
        per_period_limit_minor:
          $ref: '#/components/schemas/KesMinorUnits'
        period:
          type: string
          enum: [single_use, daily, weekly, monthly]
        category_whitelist:
          type: array
          items: { type: string }
        recipient_whitelist:
          type: array
          items: { type: string }
          description: For comms scopes — restricted set of recipient phone-token JTIs
    AuthorityStatus:
      type: string
      enum: [active, expired, revoked]
    Authority:
      type: object
      required: [id, account_uuid, agent_id, scopes, status, expires_at, created_at]
      properties:
        id:
          type: string
          pattern: '^daa_[0-9A-HJKMNP-TV-Z]{26}$'
          description: Same value as JWT `jti` claim
        account_uuid:
          $ref: '#/components/schemas/AccountUuid'
        agent_id:
          type: string
          pattern: '^agt_[0-9A-HJKMNP-TV-Z]{26}$'
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/AuthorityScope'
        status:
          $ref: '#/components/schemas/AuthorityStatus'
        token:
          type: string
          description: |
            JWT (RS256, signed by Identiti on Helpan AI's behalf). Carried in the
            `X-Delegated-Authority` header on subsequent agent dispatch calls.
            Returned only on issuance; not returned by GET endpoints.
        expires_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
          nullable: true
        revocation_reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    IssueAuthorityRequest:
      type: object
      required: [account_uuid, agent_id, scopes, ttl_seconds]
      properties:
        account_uuid: { $ref: '#/components/schemas/AccountUuid' }
        agent_id:
          type: string
          pattern: '^agt_[0-9A-HJKMNP-TV-Z]{26}$'
        scopes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AuthorityScope'
        ttl_seconds:
          type: integer
          minimum: 60
          maximum: 86400
          description: |
            Lifetime in seconds. Defaults: money-touching scopes ≤3600 (1 hour); identity-sensitive
            ≤900 (15 min); read-only ≤86400 (24 hours). Issuer enforces per-scope max.
        step_up_token:
          type: string
          description: |
            Required for high-stakes scopes (any money-touching, identity-sensitive, or
            single-use elevated-amount scope). RS256 JWT issued by Identiti per its Schema
            Appendix Amendment §A.1 with `actor.type='agent'` and `actor.agent_id` matching
            the request.
    AuthorityResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Authority'
    AuthorityListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    allOf:
                      - $ref: '#/components/schemas/Authority'
                      - description: 'Note: `token` field is omitted from list responses for security.'
                next_cursor:
                  type: string
                  nullable: true
    RevokeAuthorityRequest:
      type: object
      properties:
        reason:
          type: string
          enum:
            - user_initiated
            - operator_initiated
            - account_suspended
            - kyc_downgraded
            - cascade_user_deleted
            - cascade_consent_revoked
            - security_incident
            - other
        detail:
          type: string
    ValidateAuthorityRequest:
      type: object
      required: [token, intended_operation]
      properties:
        token:
          type: string
          description: The JWT from `X-Delegated-Authority` header
        intended_operation:
          type: string
          description: 'The operation the relying party intends to execute (e.g. `kipkiren_pay.payment.execute`, `todoku.message.send`)'
        amount_minor:
          $ref: '#/components/schemas/KesMinorUnits'
          description: For money-touching operations — the requested amount, used for limit check
    ValidateAuthorityResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              type: object
              required: [valid, status, scope_covers, within_limits]
              properties:
                valid:
                  type: boolean
                status:
                  $ref: '#/components/schemas/AuthorityStatus'
                scope_covers:
                  type: boolean
                  description: Does the authority's scope cover the `intended_operation`?
                within_limits:
                  type: boolean
                  description: Is the requested amount within the authority's amount and per-period limits?
                authority:
                  $ref: '#/components/schemas/Authority'
                  description: Authority detail (token field omitted)
                rejection_reason:
                  type: string
                  nullable: true
                  enum:
                    - null
                    - token_invalid_signature
                    - token_expired
                    - token_revoked
                    - scope_not_covered
                    - amount_exceeds_limit
                    - period_limit_exhausted
                    - account_suspended

    # ---- actions -----------------------------------------------------------
...
  responses:
    AuthHmacInvalid:
      description: HMAC signature invalid, missing, or timestamp out of tolerance
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Authentication succeeded but scope or tenant status forbids the operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RequestInvalid:
      description: Request body or parameters failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...